+++ package.js
@@ -9,6 +9,8 @@ Npm.depends({
 Package.on_use(function(api) {
 	// meteor-headers 0.0.7 was the last version to support Meteor < 0.6.5
 	api.use(['webapp', 'livedata', 'templating', 'deps'], ['client', 'server']);
+	api.use(['appcache', 'browser-policy'], 'server', { weak: true });
+
 	api.use('standard-app-packages', 'client');
 
     api.add_files('headers-common.js', ['client', 'server']);